Flytt plassering på del-skjerm
