+++ src/komponenter/header/header-simple/HeaderSimple.less
@@ -20,7 +20,7 @@
 
     &__content {
         width: 100%;
-        max-width: 1440px;
+        max-width: @desktopMaxWidth;
         height: 78px;
         display: flex;
         justify-content: space-between;
